Fix: ml의 사용자 만족도를 double로 맞춰서 계산하도록 수정

+++ src/main/java/smu/nuda/global/ml/dto/MlReviewSentimentResponse.java
@@ -7,7 +7,7 @@ public record MlReviewSentimentResponse(
         String analyzedAt
 ) {
     public record SentimentDistribution(
-            int positive,
-            int negative
+            double positive,
+            double negative
     ) {}
 }
